requirements.txt
git+https://github.com/openai/whisper.git
openai==1.65.3
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-experimental==0.3.4
langchain-openai==0.3.7
pyaudio
pydantic
#pydantic==2.11.*
#streamlit
gradio>=5.23.3
#gradio==5.23.3
langgraph==0.3.8
pandas
openpyxl
google-auth
google-api-python-client